% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssu.R
\name{cmsearch16S}
\alias{cmsearch16S}
\title{Finding 16S sequences}
\usage{
cmsearch16S(genome.file, cpu = 1, spacer = 2000, verbose = TRUE)
}
\arguments{
\item{genome.file}{A fasta-formatted file with the genome sequence(s).}

\item{cpu}{Number of CPUs to use, default is 1.}

\item{spacer}{Minimum number of bases between two distinct 16S genes on the same genome sequence.}

\item{verbose}{Logical, if \code{TRUE} progress is reported during computations (default=TRUE).}
}
\value{
A \code{data.frame} with one row for each detected 16S sequence. Similar to ORFs
this table specifies the \code{GenomeSequence}, \code{Strand},
\code{Left} and \code{Right} for each sequence, see \code{\link{orfTable}}
for details on this. There is no \code{Partial} information for 16S sequences, as their
exact start and stop varies too much.

The output from this function can be used, together with a \code{\link{Fasta}} object with the genome
to retrieve the 16S sequences by \code{\link{orfSequences}}.
}
\description{
Locating all occurrences of the 16S sequence in genomic DNA
using the Infernal software.
}
\details{
The software Infernal (Nawrocki & Eddy, 2013) must be installed and available on the system. Test 
this by typing \code{system("cmsearch -h")} in the Console, and some sensible output should be produced. 
For more details on Infernal, see http://eddylab.org/infernal/.

The module \code{cmsearch} of the Infernal software is used to search with the correlation
models for both bacteria and archaea from the Rfam database (http://rfam.xfam.org/) against the specified
genome. All significant hits (E-value below 1.0) are considered, but only unique hits are reported. In case
of multiple hits in the same genome sequence, hits are clustered by their location, using hirerachical 
clustering with single linkage. Thus, if hits are separated by less than \code{spacer} nucleotides, they are
clustered, and only the hit with the largest score in each cluster is reported. Setting \code{spacer=0} will
result in all hits being reported.

A full-length 16S sequence should be at least 
1200 bases, but (much) shorter hits may be reported. Incomplete genomes or problematic
assemblies will typically produce these shorter hits.
}
\examples{
# Using a FASTA file in this package.
# We need to uncompress it first...
\dontrun{
extdata.path <- file.path(path.package("micropan"),"extdata")
genome.file <- xzuncompress(file.path(extdata.path, "Mpneumoniae_309_genome.fsa.xz"))
#...searching for 16S sequences...
ssu.table <- cmsearch16S( genome.file )

# Inspecting the hits
print(ssu.table)

# Retrieving the sequences
gdta <- readFasta(genome.file)
fdta.16S <- orfSequences(ssu.table, gdta)

# ...and compressing the FASTA file again...
genome.file.xz <- xzcompress(genome.file)
}

}
\references{
E.P. Nawrocki and S.R. Eddy,  Infernal 1.1: 100-fold faster RNA homology searches, 
Bioinformatics 29:2933-2935 (2013).
}
\seealso{
\code{\link{orfTable}}, \code{\link{orfSequences}}.
}
\author{
Lars Snipen and Kristian Hovde Liland.
}
