% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPanGenes.R
\name{extractPanGenes}
\alias{extractPanGenes}
\title{Extracting genes of same prevalence}
\usage{
extractPanGenes(clustering, N.genomes = 1:2)
}
\arguments{
\item{clustering}{Named vector of clustering}

\item{N.genomes}{Vector specifying the number of genomes the genes should be in}
}
\value{
A table with columns
\itemize{
  \item cluster. The gene family (integer)
  \item seq_tag. The sequence tag identifying each sequence (text)
  \item N_genomes. The number of genomes in which it is found (integer)
}
}
\description{
Based on a clustering of genes, this function extracts the genes
occurring in the same number of genomes.
}
\details{
Pan-genome studies focus on the gene families obtained by some clustering,
see \code{\link{bClust}} or \code{\link{dClust}}. This function will extract the individual genes from
each genome belonging to gene families found in \code{N.genomes} genomes specified by the user.
Only the sequence tag for each gene is extracted, but the sequences can be added easily, see examples
below.
}
\examples{
# Loading clustering data in this package
data(xmpl.bclst)

# Finding genes in 5 genomes
core.tbl <- extractPanGenes(xmpl.bclst, N.genomes = 5)
#...or in a single genome
orfan.tbl <- extractPanGenes(xmpl.bclst, N.genomes = 1)

\dontrun{
# To add the sequences, assume all protein fasta files are in a folder named faa:
lapply(list.files("faa", full.names = T), readFasta) \%>\% 
  bind_rows() \%>\% 
  mutate(seq_tag = word(Header, 1, 1)) \%>\% 
  right_join(orfan.tbl, by = "seq_tag") -> orfan.tbl
  # The resulting table can be written to fasta file directly using writeFasta()
  # See also geneFamilies2fasta()
}

}
\seealso{
\code{\link{bClust}}, \code{\link{dClust}}, \code{\link{geneFamilies2fasta}}.
}
\author{
Lars Snipen.
}
