% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPanGenes.R
\name{geneFamilies2fasta}
\alias{geneFamilies2fasta}
\title{Write gene families to files}
\usage{
geneFamilies2fasta(
  pangene.tbl,
  fasta.folder,
  out.folder,
  file.ext = "fasta$|faa$|fna$|fa$",
  verbose = TRUE
)
}
\arguments{
\item{pangene.tbl}{A table listing gene families (clusters).}

\item{fasta.folder}{The folder containing the fasta files with all sequences.}

\item{out.folder}{The folder to write to.}

\item{file.ext}{The file extension to recognize the fasta files in \code{fasta.folder}.}

\item{verbose}{Logical to allow text ouput during processing}
}
\description{
Writes specified gene families to separate fasta files.
}
\details{
The argument \code{pangene.tbl} should be produced by \code{\link{extractPanGenes}} in order to
contain the columns \code{cluster}, \code{seq_tag} and \code{N_genomes} required by this function. The
files in \code{fasta.folder} must have been prepared by \code{\link{panPrep}} in order to have the proper
sequence tag information. They may contain protein sequences or DNA sequences.

If you already added the \code{Header} and \code{Sequence} information to \code{pangene.tbl} these will be
used instead of reading the files in \code{fasta.folder}, but a warning is issued.
}
\examples{
# Loading clustering data in this package
data(xmpl.bclst)

# Finding genes in 1,..,5 genomes (all genes)
all.tbl <- extractPanGenes(xmpl.bclst, N.genomes = 1:5)

\dontrun{
# All protein fasta files are in a folder named faa, and we write to the current folder:
clusters2fasta(all.tbl, fasta.folder = "faa", out.folder = ".")

# use pipe, write to folder "orfans"
extractPanGenes(xmpl.bclst, N.genomes = 1)) \%>\% 
  geneFamilies2fasta(fasta.folder = "faa", out.folder = "orfans")
}

}
\seealso{
\code{\link{extractPanGenes}}, \code{\link{writeFasta}}.
}
\author{
Lars Snipen.
}
