% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{gffSignature}
\alias{gffSignature}
\title{GFF signature text}
\usage{
gffSignature(gff.table)
}
\arguments{
\item{gff.tableat}{A \code{gff.table} (\code{data.frame}) with genomic features information.}
}
\value{
A vector of texts, one for each row in \code{gff.table}.
}
\description{
Making a signature text from \code{gff.table} data.
}
\details{
For each row in \code{gff.table} a text is created by pasting these
data together, adding some explanatory text. This function is used by \code{link{gff2fasta}}
to create the Header-lines for the \code{\link{Fasta}} object when retrieving the sequences.
}
\examples{
# See the example in the Help-file for readGFF.

}
\seealso{
\code{\link{findOrfs}}, \code{\link{gff2fasta}}.
}
\author{
Lars Snipen.
}
