% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orftools.R
\name{lorf}
\alias{lorf}
\title{Longest ORF}
\usage{
lorf(orf.table)
}
\arguments{
\item{orf.table}{A data.frame listing ORF information, see \code{\link{orfTable}}.}
}
\value{
An ORF-table with a subset of the rows of the argument \code{orf.table}. In this
ORF-table all stop coordinates are unique for each genome sequence.
}
\description{
Filtering an ORF-table to keep only the LORFs.
}
\details{
For every stop-codon there are usually mutliple possible start-codons in the same reading
frame. The LORF (Longest ORF) is defined as the longest of those ORFs sharing the same stop-codon,
i.e. the most upstream start-codon matching the stop-codon.
}
\examples{
Mer her

}
\seealso{
\code{\link{orfTable}}.
}
\author{
Lars Snipen and Kristian Hovde Liland.
}
