%% $Id: micropan-package.Rd 193 2014-09-09 08:30:27Z larssn $

\name{micropan-package}
\alias{micropan-package}
\alias{micropan}
\docType{package}
\title{
Microbial Pan-genome Analysis
}
\description{
A collection of functions for computations and visualizations of microbial pan-genomes. A pan-genome is the union of all unique gene families found among strains of a microbial species.
}
\details{
\tabular{ll}{
Package: \tab micropan\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-09\cr
License: \tab GPL-2\cr
LazyLoad: \tab true\cr
}
  See the package vignette for an introduction and how to install external software that can be used together with this package.
  
 For a guide to the use of the functions, see casestudy.pdf.
}
\author{
Lars Snipen and Kristian Hovde Liland.

Maintainer: Lars Snipen <lars.snipen@nmbu.no>
}
