% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micropan.R
\name{micropan}
\alias{micropan}
\alias{micropan-package}
\title{Microbial Pan-Genome Analysis}
\description{
A collection of functions for computations and visualizations of microbial pan-genomes.
Some of the functions make use of external software that needs to be installed on the system, see the
package vignette for more details on this.
}
\references{
Snipen, L., Liland, KH. (2015). micropan: an R-package for microbial pan-genomics. BMC Bioinformatics, 16:79.
}
\author{
Lars Snipen and Kristian Hovde Liland.

Maintainer: Lars Snipen <lars.snipen@nmbu.no>
}
