% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orftools.R
\name{orfLength}
\alias{orfLength}
\title{ORF lengths}
\usage{
orfLength(orf.table)
}
\arguments{
\item{orf.table}{A data.frame listing ORF information, see \code{\link{orfTable}}.}
}
\value{
A vector of ORF lengths, measured as the number of amino acids after translation.
}
\description{
Computes the lengths of all ORFs in an ORF-table.
}
\examples{
Mer her

}
\seealso{
\code{\link{orfTable}}.
}
\author{
Lars Snipen and Kristian Hovde Liland.
}
