% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orfs.R
\name{orfSequences}
\alias{orfSequences}
\title{Retrieving ORFs}
\usage{
orfSequences(orf.table, genome)
}
\arguments{
\item{orf.table}{A data.frame with ORF information, see \code{\link{orfTable}}.}

\item{genome}{A \code{\link{Fasta}}-object with the genome sequence(s).}
}
\value{
A \code{\link{Fasta}} object with one row for each ORF. The \code{Header} for each
ORF is a summary of the \code{orf.table} information for the same ORF.
}
\description{
Retrieving ORF sequences from a genome.
}
\details{
For each ORF listed in \code{orf.table} the sequence is retrieved from the
genome. Every \code{GenomeSequence} tag in the \code{orf.table} must match the first token
in one of the \code{genome$Header} texts.
}
\examples{
Mer her

}
\seealso{
\code{\link{orfTable}}, \code{\link{lorf}}.
}
\author{
Lars Snipen and Kristian Hovde Liland.
}
