% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orfs.R
\name{orfSignature}
\alias{orfSignature}
\title{ORF signature text}
\usage{
orfSignature(orf.table)
}
\arguments{
\item{orf.table}{A data.frame with ORF information, see \code{\link{orfTable}}.}
}
\value{
A vector of texts, one for each row in \code{orf.table}.
}
\description{
Making a signature text from ORF data.
}
\details{
For each ORF listed in \code{orf.table} a text is created by pasting these
data together, adding some explanatory text. This function is used by \code{link{orfSequences}}
to create the header-lines for the \code{\link{Fasta}} object when retrieving the ORF sequences.
}
\examples{
Mer her

}
\seealso{
\code{\link{orfTable}}, \code{\link{orfSequences}}.
}
\author{
Lars Snipen.
}
