% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orfs.R
\name{orfTable}
\alias{orfTable}
\title{Finding ORFs in genomes}
\usage{
orfTable(genome, circular = F)
}
\arguments{
\item{genome}{A \code{\link{Fasta}}-object with the genome sequence(s).}

\item{circular}{Logical indicating if the genome sequences are completed, circular sequences.}
}
\value{
A \code{data.frame} with one row for each ORF and the following columns:
\itemize{
  \item \code{GenomeSequence}. This is the first token in the \code{genome$Header} texts.
  Note that this token must be unique to each \code{genome$Header}-text.
  \item \code{Strand}. Either 1 or -1.
  \item \code{Left}. The leftmost coordinate. This is the start if the ORF is on the 1 strand, but
  the stop if it is on the -1 strand.
  \item \code{Right}. The rightmost coordinate. This is the stop if the ORF is on the 1 strand, but
  the start if it is on the -1 strand.
  \item \code{Partial}. This is 1 if the ORF in incomplete at the start of the genome sequence and -1
  it is incomplete at the end. All complete ORFs have value 0.
}

Note that the \code{Left} and \code{Right} coordinates specify the leftmost and rightmost base in the
ORF regardless of strand. If \code{Strand=1} the start-codon is in the positions (\code{Left}, 
\code{Left}+1,\code{Left}+2) and the stop-codon in (\code{Right}-2, \code{Right}-1, \code{Right}).
}
\description{
Finds all ORFs in prokaryotic genome sequences.
}
\details{
An Open Reading Frame (ORF) is defined as a subsequence starting with a  start-codon
(ATG, GTG or TTG), followed by an integer number of triplets, and ending with a stop-codon (TAA,
TGA or TAG). This function will locate all ORFs in a genome, and return a table (\code{data.frame})
of information with one row for each ORF, see below. In order to retrieve the sequences, see 
\code{\link{orfSequences}}.

Note that for any given stop-codon there are usually multiple start-codons in the same reading
frame. This function will return all, i.e. the same stop position may appear multiple times. If
you want ORFs with the most upstream start-codon only (LORFs), see \code{\link{lorf}}.

By default the genome sequences are assumed to be linear, i.e. contigs or other incomplete fragments
of a genome. In such cases there will usually be som partial ORFs at each end, i.e. ORFs where either
the start- or the stop-codon is lacking. If the supplied \code{genome} is a completed genome, with 
circular chromosome/plasmids, set the flag \code{circular=TRUE} and no partial ORFs will be listed.
You may, however, see some ORFs running across the sequence-start, with the \code{Left}-coordinate
larger than the \code{Right}-coordinate.
}
\examples{
Mer her

}
\seealso{
\code{\link{orfSequences}}, \code{\link{lorf}}.
}
\author{
Lars Snipen and Kristian Hovde Liland.
}
