% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orftools.R
\name{readOrfTable}
\alias{readOrfTable}
\alias{writeOrfTable}
\title{Reading and writing ORF tables}
\usage{
readOrfTable(in.file)
}
\arguments{
\item{in.file}{Name of file with an ORF-table.}

\item{orf.table}{A data.frame listing ORF information, see \code{\link{orfTable}}.}

\item{out.file}{Name of file.}
}
\value{
\code{readOrfTable} returns a data.frame with ORF information, see \code{\link{orfTable}}.

\code{writeOrfTable} writes an ORF-table to a text-file.
}
\description{
Reading or writing a data.frame with ORF information from/to file.
}
\details{
The \code{orf.table} must be a \code{data.frame} with the columns specified in
\code{\link{orfTable}}. These functions use \code{\link{read.table}} and \code{\link{write.table}}
to read/write such a table from/to a textfile.
}
\examples{
Mer her

}
\seealso{
\code{\link{orfTable}}, \code{\link{lorf}}.
}
\author{
Lars Snipen and Kristian Hovde Liland.
}
