% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssu.R
\name{barrnap}
\alias{barrnap}
\title{Finding rRNA genes}
\usage{
barrnap(genome.file, bacteria = TRUE, cpu = 1)
}
\arguments{
\item{genome.file}{A fasta-formatted file with the genome sequence(s).}

\item{cpu}{Number of CPUs to use, default is 1.}

\item{bacteria.}{Logical, the genome is either a bacteria (default) or an archea.}
}
\value{
A \code{gff.table} (see \code{\link{readGFF}} for details) with one row for each detected
rRNA sequence.
}
\description{
Locating all rRNA genes in genomic DNA using the barrnap software.
}
\details{
The external software barrnap is used to scan through a prokaryotic genome to detect the
rRNA genes (5S, 16S, 23S). This free software can be installed from https://github.com/tseemann/barrnap.
}
\examples{
# Using a genome FASTA file in this package.
# We need to uncompress it first...
\dontrun{
extdata.path <- file.path(path.package("micropan"),"extdata")
genome.file <- xzuncompress(file.path(extdata.path, "Mpneumoniae_309_genome.fsa.xz"))

#...searching for rRNA sequences...
ssu.table <- barrnap( genome.file )

# Inspecting the hits
print(ssu.table)

# Retrieving the sequences
genome <- readFasta(genome.file)
rRNA.fasta <- gff2fasta(ssu.table, genome)

# ...and compressing the genome FASTA file again...
genome.file.xz <- xzcompress(genome.file)
}

}
\seealso{
\code{\link{readGFF}}, \code{\link{gff2fasta}}.
}
\author{
Lars Snipen and Kristian Hovde Liland.
}
