% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssu.R
\name{barrnap}
\alias{barrnap}
\title{Finding rRNA genes}
\usage{
barrnap(genome.file, bacteria = TRUE, cpu = 1)
}
\arguments{
\item{genome.file}{A fasta-formatted file with the genome sequence(s).}

\item{cpu}{Number of CPUs to use, default is 1.}

\item{bacteria.}{Logical, the genome is either a bacteria (default) or an archea.}
}
\value{
A \code{gff.table} (see \code{\link{readGFF}} for details) with one row for each detected
rRNA sequence.
}
\description{
Locating all rRNA genes in genomic DNA using the barrnap software.
}
\details{
The external software barrnap is used to scan through a prokaryotic genome to detect the
rRNA genes (5S, 16S, 23S). This free software can be installed from https://github.com/tseemann/barrnap.
}
\note{
The barrnap software must be installed on the system for this function to work, i.e. the command
\samp{system("barrnap --help")} must be recognized as a valid command if you run it in the Console window.
}
\examples{
\dontrun{
# Using a genome file in this package.
extdata <- file.path(path.package("micropan"),"extdata")
genome.file <- "Mpneumoniae_309_genome.fsa"

# We need to uncompress it first...
xzuncompress(file.path(extdata,paste(genome.file,".xz",sep="")))

# Searching for rRNA sequences, and inspecting
gff.table <- barrnap(file.path(extdata,genome.file))
print(gff.table)

# Retrieving the sequences
genome <- readFasta(file.path(extdata,genome.file))
rRNA.fasta <- gff2fasta(ssu.table,genome)

# ...and compressing the genome FASTA file again...
xzcompress(file.path(extdata,genome.file))
}

}
\seealso{
\code{\link{readGFF}}, \code{\link{gff2fasta}}.
}
\author{
Lars Snipen and Kristian Hovde Liland.
}
