% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{gff2fasta}
\alias{gff2fasta}
\title{Retrieving sequences from genome}
\usage{
gff2fasta(gff.table, genome)
}
\arguments{
\item{gff.table}{A \code{gff.table} (\code{data.frame}) with genomic features information.}

\item{genome}{A \code{\link{Fasta}} object with the genome sequence(s).}
}
\value{
A \code{\link{Fasta}} object with one row for each row in \code{gff.table}. 
The \code{Header} for each sequence is a summary of the \code{gff.table} information in the
corresponding row of \code{gff.table}.
}
\description{
Retrieving the sequences specified in a \code{gff.table}.
}
\details{
Each row in \code{gff.table} (see \code{\link{findOrfs}}) describes a genomic feature in the \code{genome}.
This includes the \code{Seqid} indicating the genomic sequence, the coordinates \code{Start}
and \code{Stop} as well as the \code{Strand}. Every \code{Seqid} in the \code{gff.table}
must match the first token in one of the \code{genome$Header} texts.
}
\examples{
Mer her

}
\seealso{
\code{\link{findOrfs}}, \code{\link{lorfs}}.
}
\author{
Lars Snipen and Kristian Hovde Liland.
}
