% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orfs.R
\name{orfLength}
\alias{orfLength}
\title{ORF lengths}
\usage{
orfLength(gff.table)
}
\arguments{
\item{gff.table}{A \code{gff.table} (\code{data.frame}) with genomic features information.}
}
\value{
A vector of ORF lengths, measured as the number of amino acids after translation.
}
\description{
Computes the lengths of all ORFs in a \code{gff.table}.
}
\details{
See \code{\link{findOrfs}} for more on \code{gff.table}s.
}
\examples{
Mer her

}
\seealso{
\code{\link{findOrfs}}.
}
\author{
Lars Snipen and Kristian Hovde Liland.
}
