% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff.R
\name{readGFF}
\alias{readGFF}
\alias{writeGFF}
\title{Reading and writing GFF-tables}
\usage{
readGFF(in.file)
}
\arguments{
\item{in.file}{Name of file with a GFF-table.}

\item{gff.table}{A \code{gff.table} (\code{data.frame}) with genomic features information.}

\item{out.file}{Name of file.}
}
\value{
\code{readGFF} returns a \code{gff.table} with the columns described above.

\code{writeGFF} writes the supplied \code{gff.table} to a text-file.
}
\description{
Reading or writing a \code{gff.table} from/to file.
}
\details{
A \code{gff.table} is simply a \code{data.frame} with columns
adhering to the format specified by the GFF3 format, see
https://github.com/The-Sequence-Ontology/Specifications/blob/master/gff3.md for details. There is
one row for each feature.

The following columns should always be in a full \code{gff.table} of the GFF3 format:
\itemize{
  \item \code{Seqid}. A unique identifier of the genomic sequence on which the feature resides.
  \item \code{Source}. A description of the procedure that generated the feature, e.g. \code{"R-package micropan::findOrfs"}.
  \item \code{Type} The type of feature, e.g. \code{"ORF"}, \code{"16S"} etc.
  \item \code{Start}. The leftmost coordinate. This is the start if the feature is on the Sense strand, but
  the end if it is on the Antisense strand.
  \item \code{End}. The rightmost coordinate. This is the end if the feature is on the Sense strand, but
  the start if it is on the Antisense strand.
  \item \code{Score}. A numeric score (E-value, P-value) from the \code{Source}. 
  \item \code{Strand}. A \code{"+"} indicates Sense strand, a \code{"-"} Antisense.
  \item \code{Phase}. Only relevant for coding genes. the values 0, 1 or 2 indicates the reading frame, i.e. 
  the number of bases to offset the \code{Start} in order to be in the reading frame.
  \item \code{Attributes}. A single string with semicolon-separated tokens prociding additional information.
}
Any \code{NA} (missing data) is replaced by \code{"."} when writing to file. Similar, any \code{"."} read
from file is converted to \code{NA}.
}
\examples{
Mer her

}
\seealso{
\code{\link{orfTable}}, \code{\link{lorf}}.
}
\author{
Lars Snipen and Kristian Hovde Liland.
}
