% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer3.R
\name{readHmmer}
\alias{readHmmer}
\title{Reading results from a HMMER3 scan}
\usage{
readHmmer(hmmer.file, e.value = 1, use.acc = TRUE)
}
\arguments{
\item{hmmer.file}{The name of a \code{\link{hmmerScan}} result file.}

\item{e.value}{Numeric threshold, hits with E-value above this are ignored (default is 1.0).}

\item{use.acc}{Logical indicating if accession numbers should be used to identify the hits.}
}
\value{
The results are returned in a \samp{tibble} with columns \samp{Query}, \samp{Hit},
\samp{Evalue}, \samp{Score}, \samp{Start}, \samp{Stop} and \samp{Description}. \samp{Query} is the tag
identifying each query sequence. \samp{Hit} is the name or accession number for a pHMM in the database
describing patterns. The \samp{Evalue} is the \samp{ievalue} in the HMMER3 terminology. The \samp{Score}
is the HMMER3 score for the match between \samp{Query} and \samp{Hit}. The \samp{Start} and \samp{Stop}
are the positions within the \samp{Query} where the \samp{Hit} (pattern) starts and stops.
\samp{Description} is the description of the \samp{Hit}. There is one line for each hit.
}
\description{
Reading a text file produced by \code{\link{hmmerScan}}.
}
\details{
The function reads a text file produced by \code{\link{hmmerScan}}. By specifying a smaller
\samp{e.value} you filter out poorer hits, and fewer results are returned. The option \samp{use.acc}
should be turned off (FALSE) if you scan against your own database where accession numbers are lacking.
}
\examples{
# See the examples in the Help-files for dClust and hmmerScan.

}
\seealso{
\code{\link{hmmerScan}}, \code{\link{hmmerCleanOverlap}}, \code{\link{dClust}}.
}
\author{
Lars Snipen and Kristian Hovde Liland.
}
